requirements.txt
Flask==1.1.2
tqdm==4.48.0
pandas==1.1.0
numpy==1.18.5
Flask_Cors==3.0.9
Pillow==8.1.0
scikit_learn==0.24.1
